requirements.txt
asgiref==3.2.10
autopep8==1.5.3
Django==3.0.8
django-cleanup==5.0.0
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
importlib-metadata==1.7.0
Markdown==3.2.2
mysqlclient==1.4.6
Pillow==7.2.0
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
zipp==3.1.0
